fixes issue 28200: Amount field located in 'Cost' tab set as read-only
The Amount field has a read only logic based on the value of the Invoice Line field and this problem was reproducible if we tab over the Invoice Line field. After that action, the setValue() function of OBComboBoxItem was setting the value to null.

This caused that the currentValue was set to null in setItemValue function of ob-view-form, avoiding to include the invoiceline property into the currentValues of the form. As this property was not present, an undefined value was retrieved when computing the read only logic for the amount field, so the logic was not being calculated properly, causing this field to be displayed as read only.

This was an Smartclient issue in the OBComboBoxItem widget. The corresponding fix for the setValue() function has been applied to solve the issue.