Fixes issue 30971: Cancel PaymentScheduleDetails when voiding an unpaid invoice
When voiding an unpaid invoice, a dummy payment will be created related with both the actual invoice and the voided invoice.
In case there is no financial account available for invoice payment method, an error will be shown.
getFinancialAccountPaymentMethod method has been moved from AddPaymentDefaultValuesHandler to FIN_Utility, in order to use it in ProcessInvoice class.