Fixed issue 30848: The cash management is not restored after clean the cache
The feature requested is that the cashmanagement must be sent from the server
if the cahed data is not available.

The solution is to load the cash Management in the Cash Up Json and save the
information in the client side (local db).

--HG--
extra : rebase_source : fcb4f9fd7e8d25446c90e619d7f76aff3b525a8f