fixes issue 29746: unneded FIC request on record selection in some windows
This problem happens on those windows which do not have buttons on the right side of the toolbar. The EDIT request was introduced to solve issue #18711 in order to set the session variables in grid mode. This called was done through this.view.setContextInfo();. Then after fixing issue #20522 the second call in SETSESSION mode was introduced.

To avoid the call in EDIT mode we have just removed the call to this.view.setContextInfo();. This way just the call in SETSESSION mode is performed.

--HG--
extra : rebase_source : a53fcb595f9e4bc9149b612b9a1f9f0644efd404