fixes issue 28561: OB.Utilities.Number.ScientificToDecimal and leading zeros
OB.Utilities.Number.ScientificToDecimal was not working properly if the number passed as parameter had leading zeros. To fix the problem we use a regular expression to remove them: we look for at least one zero at the beginning of the string representation of the number. If found, it is deleted by replacing it with ''.