Fixed bug 31354: Wrong amtsourcedr and amtsourcecr when posting Matched Invoice
The cost is always calculated and posted in the organization's currency.
The expense is calculated and posted in the invoice's currency.

The conversion date from the cost currency to the acct. schema currency uses the shipment accounting date.
The conversion date from the invoice currency to the acct. schema currency uses the invoice accounting date.

The difference for the price variant is calculated and posted in the cost currency.

In case there is a difference in the conversion rate between the order and the invoice, the usual currency balance entry is automatically created

--HG--
extra : transplant_source : %A8%B6%F6%DA%DDx%8Di%93%E29iA%E5%F9%FCt%3FH%99