Fixes issue 31516: Empty matched document in Match Statement
The problem was that two new columns were added in aprm_matchstatement hql query in the middle of the select clause, while they were added as two new columns at the end of ad_column. As the columns order was not the same between the hql query and the ad_column and some columns did not have an alias in hql query and some others did not have an entity alias in ad_column, some column values were lost.
To fix it, we added an alias to all columns in aprm_matchstatement hql query (matchedDocument had no alias) and an entity alias to all ad_columns in aprm_matchstatement table (Transaction, bslUpdated and matchedDocument columns had no entity alias and Affinity column had a wrong entity alias).
With this, all columns of aprm_matchstatement will be correctly filled with their values in Match Statement grid.
If new columns are added to the hql query, they can go in any order in the select clause and in the ad_column, but they will need to have both an alias in the hql query and an entity alias in the ad_column, following this documentation: http://wiki.openbravo.com/wiki/How_to_create_a_HQL_Based_Table#Column_Definition