fixes issue 25012: Process definition params not work if lenght is 1000 or more
The length value was using directly as a float value to generate the js component. This caused the error parsing the component view as for example a length of 2000 result in a value of '2,000'. That comma was causing the error in the parser.

To solve the problem the float value is now converted using the 'c' built-in in the freemarker template. This built-in converts a number to string for a 'computer language', using dot as decimal separator. Also it never uses grouping separators which is the key to solve this problem.