fixes issue 29027: boolean default values for widget parameters properly set
The problem was that initially the default boolean values for widget parameters were not being set properly. This type of parameters were initially retrieved as strings, and this caused that checkboxes values were not being set correctly.

To fix the problem now the default values for boolean widget parameters are returned as boolean values from the server side.