fixes issue 30813: Show batch exception cause in SessionHandler commitAndClose
A new method, getCausingException(Throwable t), has been created in OBException to recover the underlying exception of a BatchUpdateException. This method is used in the commitAndClose() method of SessionHandler class to solve this issue.

Together with this, this new method is now used when instantating OBException with this constructor: OBException(Throwable cause) in order to retrieve the underlying exception information in this case also.