[DisplayLogicGrid] Display logic in grid
The problem was that display logic is implemented properly
when a record is being visualized/edited in form view, but
when visualizing/editing in grid view, this logic is not
taken into consideration at all.

Now, Display logic is taking into account in grid. It is
implemented it as read only logic and it is solely applied
when a record is edited in grid view but not for the rest
of records in the same grid.

Note that in "ob-view-form.js.ftl" and
"ob-pick-and-execute-view-tab.js.ftl" template files is used
"(${field.showIf} === false)" instead of "!(${field.showIf})"
to avoid "Confusing Use of !" JSLint warn.