fixes issue 31528: Can't install modules using Internet Explorer 11
The problem is that on IE 11 browser, the window.opener was lost. Now it is recovered in the same way as it was done for IE9 and Edge. This avoids the javascript error that was being fired when trying to hide the loading screen in that browser.

Together with this the cache revision number has been updated in the required files.

--HG--
extra : rebase_source : 4418baece864fcabf7c33eb5de1d49cf0ea84e61