Fixed bug 31184: Changed in order/invoice triggers to calculate Cash VAT flag
The way to calculate cash vat flag in document header was not accurate, and it failed when there was a summary level cash vat tax rate for 0%.

Now to know whether the introduced tax rate is really a cash vat tax rate (and thus to update the cash vat flag at header document accordingly), we check if exists any tax rate in the tax rate tree (or the tax rate itself if not a summary level) which is cash vat = Y and with a rate <> 0. In this case the cash vat flag at header will be automatically updated to Y, otherwise to N.
This change is done for Order and Invoice lines.