fixes issue 29259: Request with targetRecordId but without directNavigation
After pressing F5, it was not existing any mechanism to set the directNavigation parameter when refreshing the window. Besides, when closing the form view, the information of the record opened in form view (targetTabId and targetRecordId) was not being deleted from the URL.

Now, after closing the form view, that information is deleted from the URL. This way, if we refresh the window and that information is present in the URL, it means that the request must be done as direct navigation.

--HG--
extra : rebase_source : 84445e6ae547c439755ee4c1553f7558799a3dab