Fixes Issue 31115. Allow to change price of a non Invoiced Sales Order.
Previously it was fixed, but when it was done, it was allowing to change
the Sales Order Quantity when the Sales Invoice had a Goods Shipment related
and that is not right.

Changed the restriction to check changes related to prices in another
condition and allow to do this changes when the Sales Order is not
yet Invoiced, even if it has been delivered.

--HG--
extra : rebase_source : 54a1bb19566ff685562dc3d385149021871746b7