fixes issue 29025: prevents widget error in maximized view, after refreshing
After refreshing from the browser, the widget information was not being properly retrieved, causing javascript errors. Due to the lack of a infraestructure to retrieve the information for the widgets, they are not reopened after refreshing from the browser.

Now tabs maximized with OBQueryListView and OBCalendarWidgetView are not recorded in the history (URL) this way they will not be reopened.

--HG--
extra : rebase_source : c817eafb205c780334aa9c23a0dea6dc3f28f157