Fixed bug 31107: Commission for full paid documents split in several payments
For Full paid documents, the query will search for the invoices that are fully paid where the last payment belongs to the selected period.
So the query has been transformed to get all the PSD for the fully paid invoices if it exists at least one PSD paid in the given period and when all the PSD for the invoice are before the "to date" of the given period.