fixes issue 31566: Error when closing window having focus on a yes/no filter
When closing the tab having the focus on a OBYesNoItem, the handleEditorExit() method of this class is executed, causing an unexpected javascript error on the getCriteriaFieldName() method, because is not possible to reference the form as it is being destroyed.

It does not make sense to execute handleEditorExit() when closing (destroying) the window. So to avoid this execution we introduce a new flag for the items (isBeingDestroyed) to mark that they are being destroyed. Thus, it is used for OBYesNoItem to identify whether standard handleEditorExit() flow must be executed or not.

--HG--
extra : rebase_source : b25d8b05ecee0c2cc10fb682c6a2367e67fcbb1a