fixes issue 31578: alert fetches with where parameter
In order to avoid the fetching of the where parameter we generate the where clause inside the datasource. A new datasource has been created, ADAlertDatasourceService, to support the requests from the Alert Management window.

In addition, it has been fixed a problem in the Alert Rule drop-down filter of the grids in the Alert Management window which was displaying alert rules from all the grids in the window instead of those displayed just on a particular grid.