fixes issue 27240: On Change Function in Process Definition non editable fields
The handleItemChange method is the method in charge of execute the on change function on onblur. This event is not fired when setting the value of a field programmatically. To solve this, a new method has been introduced in the API: setValueAndHandleItemChange() which invokes handleItemChange() after setting the value of a the field.

This method will be available automatically just for those Process Definition fields with an on change function assigned to them through the application dictionary.

--HG--
extra : rebase_source : 4757c2396471b1381fecc51e856a6eb7ce104d75