Fixes issue 30879 : Show More Button by default in Payment Method if there are more than 5 Payment Types
If Payment types are more than 5 and default payment type(Cash) is not present in first four,
then More Button label will be displayed as "More... (Cash)", if we click it, it will display the Payment Type Popup.

If Selected Payment Type is not present in the Side Button Payment Type List, then
More Button Label will display as "More... (**)", ** = Selected Payment Type.