Fixed issue 26349: Robustness:Improve handling of disabling triggers.
It has been improved the handling of disabling triggers by throwing
exceptions in enable() and disable() methods instead of error logs.

Now an exception is thrown if there is a problem in enable() and
disable() methods.

Besides it has been added a change to improve where clear() method
is invoked.