fixes issue 31683: Some values are not shown in XLS format using MS Office
The detect cell type feature was not working properly for BigInteger values in Office after the Jasper Library Upgrade. A personalized format was applied causing the data not be shown when opening the report.

To solve this problem the detect cell type feature has been disabled. This way, by default, all the values in the XLS will be exported as strings.

This default configuration can be overriden at report level. As an example, the General Ledger Report and Trial Balance report have been updated properly. With this configuration, those reports are now using number formatting for the cells showing the BigDecimal values.