Fixed issue 30771: CalculateTaxes should not be executed when the receipt is being closed
The problem was that, while the recipt is being closed, the recipt could change
in the process.

The proposed solution is to change the places and the way the POS calculates
the promotions, taxes and the gross of the receipt.

With these objective, it's created the function calculateReceipt. The extra
calculateTaxes, applyPromotions and calculateGross functions have been deleted
from some files.

--HG--
extra : rebase_source : 49ea22d2afe233f9016643fe95114a06d893e552