[pi-field-filt-sort] Change of the two of boolean variables.
Two new boolean variables have been created for controlling if the canSort and
canFilter properties can be overwritten or not, have been created and set in the
proper place. There were other boolean varibles but their name was not the best
for the use they have. The place was not the correct neither.