fixes issue 31722: a date used to filter on a grid dissapears after enter it
After solving issue 29554 the values of the date filters are completely removed when invoking the clearValue() method. For this reason, in order to prevent the clearing of the date filter it is necessary to call setValue().

This was already controlled when choosing the dates from the date pickers, but not after entering the date with the tab or enter keys or when using the 'Use as Filter' menu option.

Now this cases are controlled also, by calling setValue() on the setSingleDateValue() method.