Related to issue 31701: code review improvements
When reactivating an order, do not consider payments in Awaiting Payment status as payments really linked to the order. This is a way to ensure reactivation of orders already linked to a voided invoice is possible, because since #30971 a 0 amount payment is automatically when voiding an invoice.
So the flow in that case will be to manually reactivate that payment, and then reactivate the order.

Updated copyright year.