fixed issue 31749: The warning message is not correct.
When a  warning message had to be shown when saving a record, it was not the proper message.
It showed "Saving failed" text appended to the original warning message.
That was not correct because it was possible to save the record.
To fix this the "setMessage" function of the "ob-messagebar.js" has been changed.
Now, it appends the "Saving failed" message only if the message type is an error.