Fixed issue 30771: CalculateTaxes should not be executed when the receipt is being closed
If you create a ticket with a pricelist with the taxes not included and you
reload the POS before you pay it, when you try to pay it after the reload a
javascript error is show.

The problem is that the receipt save a value as big decimal, and when is
retrieve from the local database the value is no longer a number, it's an
object.

The solution is to save the value of discountedprice as a number instead
of big decimal.

--HG--
extra : rebase_source : 3940ac25241bfe06fa1021533a93f312013bd973