Fixed bug 31697: code review improvements
In case of removing a gl item from a transaction, the part of the transaction's description related to the GL Item (and everything else afterwards) will be deleted. The part of the transaction's description before the GL Item description
 will be always kept.
The same behavior is applied when adding a new transaction from the Financial Account window and from the Match Statement --> Add transaction process

In case of removing a payment, the behavior is similar to the GL Item described above. However, since the payment description doesn't follow a constant structure (like in the case of GL Item's description), it's very hard to detect the description on the fly. That means that the flow will only work just after saving the transaction or in the first change after loading the transaction, which are the common flows. If the user selects different payments without saving the record, the system will append the different payment's description to the transaction.

The getFinAccTransactionDescription() method:
1. won't append multiple times the same description.
2. will add an extra empty line only when it's something to be appended
3. will work fine when the transaction already has a description with multiple lines
4. control any possible exception and returns the text to be append (if not blank), or the original description

Applied jsbeautifier to ob-aprm-addTransaction.js