Fixed bug 31760: Some labels in Add Payment can't be translated
Solution partially based on ALO's one.
The query for translation was wrongly using the element TRL ID instead of the Application Element ID.
Avoid querying the database when language is en_US

This piece of code is executed from Purchase Invoice, Sales Invoice, Payment In, Payment Out, Financial Account | Transaction, Financial Account | Match Statement. However, only in the Financial Account flows seem to be really useful, so we should think about not running it in the other places.