Fixed issue 31753: Instances should automatically start using the discounts new flow starting 16Q1
Starting from 16Q1 the old flow of discounts is not supported.

Some changes are necessary:
* The preference OBPOS_discount.newFlow must disappear.
* All modulescripts related to this preference must be deleted.

Now, the preference doesn't exists in the backend, and there is no modulescript
using it. This preference is set to true when the terminal properties are been
loaded.

It's planned that in 16Q2, all checks of this preference will be deleted in
the client side.