fixes issue 31647: error after press ESC on empty grid of a process definition
The pneSelectionUpdated() function assumes that the actualRecord parameter is not null. If this method is invoked when a null or undefined value an error is thrown when trying to access to the id property of actualRecord. To prevent this error now we check if actualRecord is different from null and undefined before invoking pneSelectionUpdated() on the selectRecord() method of OBPickAndExecuteGrid