fixes issue 31331: When add new lines in sales order the product appears filled
Two different flows were resulting in this problem. From one side, if we clicked directly on the new record button quickly after filling the combo value it could happen that the new form was created before the filter operation of the combo finished. To overcome this problem now a temporary id is generated every time we edit/create a new record in form. This id is checked once the combo filter operation ends, to identify if the form has changed. In that case, the flow to be executed after the filter operations is cancelled.

From other side, the same proble could happen if we clicked on the new record button quickly after clicking on the save record button. In this case, it could happen that after clearing the values a wrong entry was retrieved for the combo using the mapValueToDisplay() function. To solve this problem, now in the setEntries() function we include the value field information. This helps to retrieve the correct value on the described circumnstances.

Together with this, a new utility function has been created in order to have a common function to generate 'temporary' ids.