fixes issue 31454: Can not filter by Attribute Set in Return From Customer P&E
We are making use of the existing creatingJoinsInWhereClauseIsPrevented flag, which is true for HQLDataSourceService, to avoid the join resolving in the computeLeftWhereClauseForIdentifier() method of the AdvancedQueryBuilder class.

Before this fix, a new alias was being included in the where clause but it was never defined before, causing the generation of a wrong query.