fixes issue 31884: can not move tree nodes in classic mode
When moving tree nodes using the classic mode there is a point where the process tries to recover the tree type. This information is stored in the ADTable entity which can not be accessed by non system roles. To prevent this error now the part where the tree type is recovered is executed using the DAL admin mode.