Fixed bug 31873:Multicurrency accounting posting partial invoice with prepayment
The selectPrepayments query, used only for foreign currency accounting schemas, was wrongly taken into account all the payment details linked to the order.
In case of an invoice linked to a prepaid order, when the invoice's amount is lower than the order's amount, the system automatically creates 2 payment details both linked to the order, and one of them linked to the invoice too.

After the fix, the query only takes into account the payment details linked to both the order and the invoice being posted.