Fixed issue 31827: 401 errors
The problem was that the session could be killed before the synchronization
process has finished.

The solution (specially in the cashup) is to wait to the end of the
synchronization process before the confimation popup (or the POS
hardware manager popup error) is shown.

--HG--
extra : rebase_source : 9d52fb505b455548bf4b93e0de282793d6fd714b