fixes issue 31904: 'Filter on change' option set to N does not work
The text items were not accessing properly to the 'filter on change' property. For P&E grids an additional problem was happening: the value of a variable called actOnKeypress used to disable/enbale the filtering on change was being overriden by the evaluateDisplayLogicForGridColumns() function. Also, the filter when pressing enter was not being working neither because the editFormDefaults values were not being used properly in P&E.

--HG--
extra : rebase_source : 8e9bb871671326a994604e85ca9a6b604494ebd5