[pi-security-hqlinjection] Don't send _where param in Pick and Execute windows
The _where parameter is not sent in the pick and execute windows' datasource.
Now a boolean parameter is sent, "isFilterApplied", this parameter is true when
there is a filter applied in a pick and execute window and it is false when there
is not a filter applied or it has been cleaned up.