fixes issue 31961: Incorrect posting data using postgresql jdbc connection pool
There was a problem in the binary transfer of Dates performed by the postgresql library version 9.4-1201. This problem is not happening on version 9.4-1207 because when the queries are executed it now includes a mechanism to identify if the fields of the query need to update their format to binary.

For this reason, to fix the problem the postgresql jdbc driver has been upgraded to version 9.4-1207