Fixes issue 31965: Enables loading the model without standardizing PLSQL code
The platform interface has now a method, loadModelFromDatabase(ExcludeFilter filter, boolean doPlSqlStandardization), that accepts a parameter to specify whether the PLSQL code standardization should be done when the database model is loaded. This method would be useful for those who want to load the model but that don't need to do anything related with the PLSQL code (for instance if they just want to use the model to disable/enable foreign keys and constraints).

Adds also tests to check that it is possible to load the database model with and without doing the PLSQL code standardization.