[pi-security-hqlinjection] Create the proper code in the backend
The way of handling the where and filter clauses now it is different. In the DefaultJsonDataService.java, a
protected method has been created, to be able to invoke it in the same package.
This method is the one which creates the query that, before, was sent in the _where parameter of the datasouce.