fixes issue 31901: can't edit decimal value with observer after manual edition
The decimal fields were using a property called textualValue to prevent rounding errors. This is because when instantiating a BigDecimal using a double value, it converts the entire precision of the double. Then, on the server, the textual value was translated back to a number using BigDecimal avoiding the rounding problem.

In this case the problem was caused because the textualValue was not being synchronized properly with the current value. To solve the problem, the textualValue property is not longer used and the textual conversion is done in the server: using the BigDecimal.valueOf method the double value is converted to a string internally, and finally that value is converted into a BigDecimal.

Thus, we do not need to worry about keeping textualValue synchronized and the double value is rounded properly using BigDecimal.valueOf()