fixes issue 32007: combo/selector with a callout can cause a double FIC request
If a form redrawal was fired after selecting a value of a combo/selector with a callout, a double request to the datasource was being performed due to asynchrony problems. An example of this scenario is when after selecting a value of a combo/selector with a callout, the form needs to be scrolled (overflow) because new fields are displayed in the form (for example because this fields have a display logic).

This kind of scenarios are now detected in the change event of the OBSelectorItem class, when the value to be updated is null but at the same time the pick list already contains a selected value.