fixes issue 31180: openDirectTab action now works if the tab is already opened
Now the 'openDirectTab' response action takes into account if the 'Allow Multiple Tabs of a Window' preference is enabled. In that case when this action tries to open a tab which is already opened it opens a new tab in the last position instead of refreshing the old one.

--HG--
extra : amend_source : 5e3221d37d191330c2f3647e208991ed082da79b