fixes issue 31180: openDirectTab action now works if the tab is already opened
When the opening tab was already opened, the openDirectTab method was not waiting properly to the new tab be refreshed/opened. This was caused because the callout of OB.Utilities.openDirectTab method was invoked after the line that checks if the action needs to wait.
Now if the tab is already opened the action is called again with a delay of 100ms, giving OB.Utilities.openDirectTab enough time to execute its callout.