[pi-security-hqlinjection] Code review fix for Alert Management window
Before, the where clause was being overwritten for every manual dataosurces.
This has been fixed by creating a new method which tells if a datasource is manual
or not and if it is manual, now, the where parametes is added the same way as it is
for the selector items.