Fixes issue 31712: Foreign keys and constraints can be disabled per table
Before this change the Platform class only offered methods to disable all the constraints and foreign keys of the whole database. Now it is possible to do it also on a table to table basis.

Changes done:
- Platform.java: Added public methods disableAllFkForTable, enableAllFkForTable, disableCheckConstraintsForTable, enableCheckConstraintsForTable
- PlatformImplBase.java: Provides a functional default implementation for disableCheckConstraintsForTable and enableCheckConstraintsForTable, possible because the .disableAllChecks(table) was already implemented. Also provides an implementation for disableAllFkForTable and enableAllFkForTable that just returns throw new DatabaseOperationException("Error: Operation not supported"), because it needs to be implemented in each specific platform.
- Oracle8Platform and PostgreSqlPlatform: Implementations for disableAllFkForTable and enableAllFkForTable.

Some tests have been added to check that the foreign keys and constraints can be disabled and enabled for the whole database and also for a specific table.

--HG--
extra : rebase_source : 9e83e849ebd4a7d92c2b73f2b9be68bf942d6538