Fixes Issue 32038: "Price Adjustment" promotion type is applied to a line only if the order/invoice date falls into the promotion period.
The HQL query was tuned to take into account the starting and ending date of the
promotion when selecting a product for a line in a purchase order/invoice.
The line only gets the promotion if the period encloses the order/invoice date,
start and end date included.
The line hql += "and (endingDate is null or endingDate>:date) "; was replaced with

hql += "and (endingDate is null or endingDate>=:date) ";
hql += "and (startingDate is null or startingDate<=:date) ";

--HG--
extra : transplant_source : %FF%C2%28%D1%C5%F7%86%D6%96yKM%05%B0%E4%ADj%20M%BD