fixes issue 32092: hidden fields space not restored after reopening form view
The alwaysTakeSpace property of the form items was only being set after opening the form view the first time. This property is used to keep space for the displayed field (including the hidden ones by display logic). This property is being set to false when closing form view, and then, never restored again.

Now this property is set every time the UI switches into the form view.