Fixed issue 31913: link physical device does not work when the role has not got permission to Terminal window
If a user doesn't have permission to the Terminal window and the terminal
authentication is active,  when the user log in the WEB POS, a error is shown
in the openbravo log.

The solution is to grant access.

--HG--
extra : rebase_source : ae294133ed0b2ca7abea6fe6ba2eb137988c226f