Fixes issue 32078: Reserved quantity is not correct
In ManageReservationActionHandler, when clicking Done in Manage Reservation P&E, it was first reserving what the user was typing in the P&E, and if it was not processed the reservation, it was processing it. As the M_RESERVATION_POST tries always to reserve as much as possible, it was overriding the previously set amounts. In order to prevent this, now first the reservation is being processed and after that the values typed in the UI are being set to the reservation.
In case no stock reservation is created by m_reservation_post when processing the reservation, res_status will be Draft. To be sure reservation is set as completed, force status after call ReservationUtils.processReserve method.
Reservation will only be kept as draft in case it has no lines after execute manageReservedStockLines.