related to issue 32091: code review improvements
After sorting the grid without having any filter applied, the selection was not being kept on top. To solve this, now we check if there is a criteria present in the request and also if there are selected records. In that case, a new criteria is included in the request with the ids of the selected records.

Besides, two methods are deleted from AddPaymentOrderInvoicesTransformer as they are no longer needed. And a new check is added when retrieving the selected records in the backend, in order to avoid a possible JSONException.

--HG--
extra : rebase_source : e2a0e5c5e997eb9eec93d3ff6b44920780114d16