Fixed bug 32100: Reactivation control of a prepaid cash vat invoice
When an invoice is linked to a payment, the system currently avoids to reactivate it. You first need to reactivate the payment, so the cash vat information is deleted, and then you can reactivate the invoice.
However for prepaid invoices, the system does allow to reactivate the invoice, but the cash vat information is wrongly kept.

Now we delete all the cash vat info records linked to a prepayment for this invoice at invoice reactivation time. When the invoice is latter on processed, the cash vat records will be calculated again.