fixes issue 32202: Error in add payment when editing the amount of a record
The entered value is being set on the changed method for number items. This is done in order to prevent placing the decimal separator in a wrong place. This problem only affects to standard windows.

To solve the problem we only set the number item value in the changed method when we are not in a pick and edit window but in a standard window.

--HG--
extra : rebase_source : fe79e245eb87b057e2957750c7a1df5a5ecc7a95