[pi-security-hqlinjection] Changed the name of a method
The method named addTransactionalFilter is now called getFilterClause.
This is because before there exists a method called getFilterClause but it
only did a verification that has been moved to the "addTransactionalFilter" method.
So, now, the addTransacationalFilter method is the one which returns the filter clause.
Sometimes it has a transactional filter too.