fixes issue 32231: wrong number comparison when checking field access of a role
When checking the access of a numeric field, if it is an integer, the client sends the value without decimal places. This value is compared with the current value stored in the database which already stores the value with decimals. So, the comparison of the string values determines that they are different as for example we are comparing '100' with '100.00'.

To solve the problem, for this type of fields now the numeric value is compared instead of using its string representation.

--HG--
extra : rebase_source : 3bf22c3fcd7927d3acba0ec9e10b0f84c8bcf795