[pi-security-hqlinjection] Changed a logic added before
Before, a logic in the AdvancedQueryBuilder was added to handle the behavior of
the selectors. The tested selector was the "Action Regarding Document" of the "Add Details"
popup of the "Payment In window". With this fix, the selector worked fine and showed only the
expected value. But this fix was breaking the behavior of the FK dorpdaows, for example the dropdown
of the "Produc"t selector popup of the "lines" subtab of the "Sales Order" window was broken with this fix.
So, the fix has been reverted.