fixes issue 32193: Fix selection in P&E grids having filter on change disabled
Having filter on change disabled, the filtering is not performed until enter is pressed or the focus is moved out of the text filter. In this second case, there could be problems if the focus is lost when checking a pick and edit grid record: the filtering request is performed and at the same time the actions related to the record selection are fired. This can cause the selection actions not work as expected.

This problem also affects to numeric and date filters, because they always perform the filtering after losing the focus.

To avoid this problem a new mechanism has been included for pick and edit grids, the blur method for the filter elements is overridden to prevent the selection immediately after losing the focus.