fixes issue 32287: Selection can be lost in P&E grids under some circumstances
If we refresh the grid, after selecting a record in a page different from the first one, Smartclient internally prepares the request to retrieve the records surrounding the selected one, in order to try to keep the position of the scroll.

This behavior in P&E can cause the losing of the selected records, because they are always loaded on the first page. In order to prevent this, we detect this situation in P&E grids before performing the request, and in that case we adapt it to force the retrieval of the first page which contains all the selected records.

--HG--
extra : amend_source : 95abc587a650406846662f030d4cb1db22aef59d