Fixes issue 32240: Read only doesn't work on some columns in grid view
The problem is caused because it was not take into account read only when other
logic (e.g. display logic) was being applied in an editing grid view.

In order to prevent this, it take into account when a field is checks as read
only. In this case, the field is show properly.